flex sync working for all except create
